+++ contracts/src/proxy/utils/mod.rs
@@ -1,0 +1,4 @@
+//! This module contains utility contracts for the proxy pattern.
+
+pub mod erc1822;
+pub mod uups_upgradeable;