+++ examples/ownable/src/lib.rs
@@ -3,7 +3,7 @@ extern crate alloc;
 
 use alloc::vec::Vec;
 
-use alloy_primitives::{Address, U256};
+use alloy_primitives::{Address, B256, U256};
 use openzeppelin_stylus::{
     access::ownable::{self, IOwnable, Ownable},
     token::erc20::{self, Erc20, IErc20},
@@ -63,6 +63,13 @@ impl OwnableExample {
     #[constructor]
     fn constructor(&mut self, initial_owner: Address) -> Result<(), Error> {
         Ok(self.ownable.constructor(initial_owner)?)
+    }
+
+    // Dummy function for some other E2E tests.
+    // e.g. UUPS Proxy example: `upgrade_to_invalid_proxiable_uuid_reverts`.
+    #[selector(name = "proxiableUUID")]
+    fn proxiable_uuid(&self) -> Result<B256, Vec<u8>> {
+        Ok(B256::ZERO)
     }
 }